+++ src/main/java/no/uib/inf101/sem2/gameEngine/view/raycaster/LinearMath/Plane.java
@@ -1,3 +1,3 @@
 package no.uib.inf101.sem2.gameEngine.view.raycaster.LinearMath;
 
-public record Plane(Vector normal, double dist) {}
+public record Plane(Vector normal, float dist) {}
